Move LabeledCheckbox into modifiers UI package
