+++ src/main/java/thestonedturtle/bankedexperience/data/modifiers/ui/ModifierPanel.java
@@ -29,7 +29,6 @@ import javax.swing.JComponent;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
-import thestonedturtle.bankedexperience.components.LabeledCheckbox;
 import thestonedturtle.bankedexperience.data.modifiers.Modifier;
 import thestonedturtle.bankedexperience.data.modifiers.ModifierComponent;
